% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LI600_analyze_flash.R
\name{LI600_analyze_flash}
\alias{LI600_analyze_flash}
\title{A function to read and plot LI600 flash data}
\usage{
LI600_analyze_flash(file_loc, main_title)
}
\arguments{
\item{file_loc}{file containing flash data.  Same format as \link[utils]{read.csv}.}

\item{main_title}{main plot title.}
}
\description{
This is a convenience function that combines \link{LI600_read_flash} and \link{LI600_plot_flash}.
}
