% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LI600_plot_flash.R
\name{LI600_plot_flash}
\alias{LI600_plot_flash}
\title{A function to plot LI600 flash data}
\usage{
LI600_plot_flash(flash_df, main_title = NULL)
}
\arguments{
\item{flash_df}{flash data frame.  Data frame should only contain data from a single flash.}

\item{main_title}{main plot title.}
}
\description{
This function will plot the flash data from the LI600 to aid in ensuring
proper flash parameters are used.  It will plot the full plot and
top 5\%, 2.5\%, and 1\%.
}
