% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LI600_read_data.R
\name{LI600_read_data}
\alias{LI600_read_data}
\title{A function to read LI-600 data files.}
\usage{
LI600_read_data(file_loc)
}
\arguments{
\item{file_loc}{file containing flash data.  Same format as \link[utils]{read.csv}.}
}
\description{
This function will read LI-600 main data files and remove secondary headers.
See \link{LI600_read_flash} for reading flash files.
}
